smartsdlc/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart
python-dotenv
pymupdf==1.24.5
requests
